Added StringReverser
Added a Java source file which has a public method which takes a String
variable and returns it with its characters in reverse order.

The main method does this for a String fed into it in the arguments in a
CLI, or otherwise a String of the alphabet; it prints the original form
of the String, then the reversed String.